pyproject.toml
[project]
name = "galera-volei"
version = "1.0.0"
description = "API REST para gerenciar partidas de vôlei com autenticação JWT"
authors = [
    {name = "Nazir Reis"}
]
requires-python = ">=3.9,<3.13"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "sqlalchemy>=2.0.20,<2.1.0",
    "alembic>=1.12.1,<1.14.0",
    "psycopg2-binary>=2.9.7,<2.10.0",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "bcrypt==4.0.1",
    "python-multipart>=0.0.6,<0.1.0",
    "email-validator>=2.1.1,<3.0.0",
    "pydantic[email]>=2.5.0,<2.6.0",
    "pydantic-settings>=2.0.0,<2.2.0",
    "python-dotenv>=1.0.0,<1.1.0",
    "pytest>=7.4.0,<8.0.0",
    "httpx>=0.24.1,<0.26.0",
    "requests>=2.32.5",
]

[project.optional-dependencies]
test = [
    "pytest",
    "httpx==0.24.1",
]
dev = [
    "pytest",
    "httpx==0.24.1",
]
